feat: rename 'updater' to 'update' and 'self-update' to 'update self'
