+++ pkg/commands/internal/update/main.go
@@ -4,7 +4,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-package updater
+package update
 
 import (
 	"fmt"
